requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.6.0
    # via
    #   -r requirements/base.txt
    #   django
attrs==22.2.0
    # via pytest
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   pynacl
click==8.1.3
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==7.2.2
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-model-utils
    #   edx-django-utils
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-model-utils==4.3.1
    # via -r requirements/base.txt
django-waffle==3.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
edx-django-utils==5.3.0
    # via -r requirements/base.txt
edx-opaque-keys @ git+https://github.com/openedx/opaque-keys.git@2.3.0
    # via -r requirements/test.in
exceptiongroup==1.1.1
    # via pytest
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via code-annotations
markupsafe==2.1.2
    # via jinja2
newrelic==8.8.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==23.0
    # via pytest
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.0.0
    # via pytest
psutil==5.9.4
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pymongo==3.13.0
    # via edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==7.2.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-slugify==8.0.1
    # via code-annotations
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==6.0
    # via code-annotations
sqlparse==0.4.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.0.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
